# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Empresas",
        "description": "",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/",
            "description": "Entorno de Producción (Consume Créditos)"
        },
        {
            "url": "/api/sandbox/v1",
            "description": "Entorno Sandbox de Pruebas (Gratis. Usa CIF: A15075062)"
        }
    ],
    "paths": {
...
        "/api/v1/companies/search": {
            "get": {
                "tags": [
                    "1. Plan Free"
                ],
                "summary": "Búsqueda de Empresas",
                "description": "Busca empresas por nombre o CIF. Permite búsqueda exacta o paginada. **Coste:** 1 llamada de tu cuota mensual (plan suscripción) o 1 crédito del monedero (bono prepago). Las respuestas con error (400, 404, etc.) no consumen cuota ni créditos.",
                "operationId": "7505c6f7c5eb0d2af39a3177326b5617",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Nombre o CIF de la empresa a buscar",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "multiple",
                        "in": "query",
                        "description": "Si es true, devuelve una lista paginada de coincidencias. Si es false, devuelve la mejor coincidencia.",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "default": false
                        }
                    },
                    {
                        "name": "cursor",
                        "in": "query",
                        "description": "Cursor opaco para paginación (recomendado para integraciones nuevas). Si se proporciona, sobrescribe el parámetro `page`.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Número máximo de resultados por página (solo si multiple=true). Máximo 100.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página (solo si multiple=true).",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados de la búsqueda",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "ApiKeyAuth": {
                "type": "apiKey",
                "description": "Introduce tu API Key privada aquí para autenticarte y poder probar los endpoints en tiempo real.",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "ApiKeyAuth": []
        }
    ],
...
}